fix(teams): implement proper leave and token regeneration
