+++ src/pages/teams.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../layouts/Layout.astro';
-import TeamOverview from '../components/app/pages/Team.svelte';
+import TeamOverview from '../components/app/pages/Teams.svelte';
 import Header from '../components/app/lib/Header.svelte';
 import { lucia } from '../lib/lucia';
 import { DUMMY_SESSION } from '../lib/helpers';
